// src/engines/particle-orb/particleOrb.types.ts
export type ParticleOrbConfig = {
  orbitSpeed: number;
  energyResponse: number;
  particleCount: number;
  glowIntensity: number;
  spaceScale: number;
  colorSaturation: number;
  sparkleDensity: number;
  warmth: number;
  primaryColor: string;
  secondaryColor: string;
  showTitle: boolean;
};